+++ pkg/registry/motif/includes.go
@@ -1,6 +1,8 @@
 package motif
 
 import (
+	"fmt"
+
 	orktypes "github.com/orkspace/orkestra/pkg/types"
 )
 
@@ -20,6 +22,20 @@ func expandIncludes(m *orktypes.Motif, dir string) error {
 		}
 		if err := orktypes.ExpandMutationInclude(m.Admission.Mutation, dir); err != nil {
 			return err
+		}
+		if m.Admission.Validation != nil {
+			var err error
+			m.Admission.Validation.External, err = orktypes.ExpandExternalCalls(m.Admission.Validation.External, dir)
+			if err != nil {
+				return fmt.Errorf("admission.validation.external: %w", err)
+			}
+		}
+		if m.Admission.Mutation != nil {
+			var err error
+			m.Admission.Mutation.External, err = orktypes.ExpandExternalCalls(m.Admission.Mutation.External, dir)
+			if err != nil {
+				return fmt.Errorf("admission.mutation.external: %w", err)
+			}
 		}
 	}
 	return nil